Coc-Admin-App. List component added

+++ Apps/CoC-Admin-App/src/app/components/login/login.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { FormControl, Validators } from '@angular/forms';
+import { FormControl, Validators } from '@angular/forms'
 
-import { AngularFireAuth } from '@angular/fire/auth';
+import { FirebaseService } from '../../services/firebase.service';
 
 @Component({
   selector: 'app-login',
@@ -13,8 +13,9 @@ export class LoginComponent implements OnInit{
   email:FormControl;
   emailSent:boolean;
   emailError:boolean;
+  aux:string;
 
-  constructor(private fireAuth:AngularFireAuth, private router:Router) {
+  constructor(private router:Router, private firebase:FirebaseService) {
     this.email = new FormControl('', [Validators.required, Validators.email]);
     this.emailSent = false;
     this.emailError = false;
@@ -22,57 +23,36 @@ export class LoginComponent implements OnInit{
 
   ngOnInit() {
     const url = this.router.url;
-    if(url.includes('signIn')){
-      this.confirmSignIn(url);
+    if(url.includes('signIn')){ //Esto quiere decir que el componente se ha inicializado gracias al link de autenticación
+      this.firebase.confirmSignIn(url).then(response=>{ //Si la confirmación tiene éxito, navegar al listado de requests
+        window.localStorage.removeItem('emailCoCAdmin');
+        this.router.navigate(['/list']);
+      }).catch(err=>{
+        console.log(err);
+      });
     }
   }
 
   sendEmailLink(){
     if(this.email.valid){
-      this.emailError = false;
-
-      //Prepare things to send email
-      let actionCodeSettings = {
-        url: 'http://localhost:4200/login',
-        handleCodeInApp: true
-      };
-      //Send mail
-      this.fireAuth.auth.sendSignInLinkToEmail(this.email.value, actionCodeSettings).then(response=>{
-        console.log(response);
-        //Store email in local storage
-        window.localStorage.setItem('emailCoCAdmin', this.email.value);
-        this.emailSent = true;
-      }).catch(err=>{
-        console.log(err);
-      });
-      
-    }else{
-      this.emailError = true;
-    }
-    
-  }
-
-  confirmSignIn(url:string){
-    if(this.fireAuth.auth.isSignInWithEmailLink(url)){
-      let email = window.localStorage.getItem('emailCoCAdmin');
-
-      //If missing email, prompt user for it
-      if (!email){
-        email = window.prompt('Please provide your email for confirmation');
-      }
-
-      //Signin user and remove the email from local storage
-      this.fireAuth.auth.signInWithEmailLink(email, url).then(response=>{
-        console.log(response);
-        window.localStorage.removeItem('emailCoCAdmin');
-      }).catch(err=>{
-        console.log(err);
+      //Check if the email exists in the collection of admins
+      this.firebase.getAdmin(this.email.value).subscribe(response=>{
+        if(response !== undefined){ //Email exists
+          //Send link to admin email
+          this.firebase.sendEmailLink(this.email.value).then(response=>{
+            //Store email in local storage
+            window.localStorage.setItem('emailCoCAdmin', this.email.value);
+            this.emailError = false;
+            this.emailSent = true;
+          }).catch(err=>{
+            console.log(err);
+          });
+        }else{ //Email does not exist
+          this.aux = this.email.value;
+          this.emailError = true;
+        }
       });
     }
-  }
-
-  logout(){
-    return this.fireAuth.auth.signOut();
   }
 
 }
